Code Guidelines - Notes about method placement
Add some detailed notes on method placement, with link to Clean Code Chapter. 

For the most part I do think the existing style in TripleA mostly matches the strategy described in Clean Code. We do have a lot of examples of private methods written above where they are first used, but often that context is a really long method and so we put the method there more because it was close, rather than it being a good place for it. In the more cleaned up/structured code in triplea we do not see examples of this. Some examples are the 'ui widgets' that follow what I think is bad pattern (series of 'void' method calls), but nonetheless demonstrates the point by at a high level by defining methods in the order at which they are called, eg:
```
constructor() {
   createWidgets();
   layoutWidgets();
   activateWidgets();
}

createWidgets() {

}
layoutWidgets() {

}

activateWidgets() {

}
```