README - add content back that was moved to the website docs repo.
Previously this info was here to begin with, so it's been moving around.
What it is - the basic getting set up on the website. Has all to do with this repo. Instead of this info living with the game engine, it is more pertinent here. The game engine docs now link here isntead of the other way round. This should make it a bit easier if you just want to do website work to find this repo, see the instructions and get running.