Remove about page (#287)
- History content from that page has been moved to forum post: https://forums.triplea-game.org/topic/523/an-early-triplea-history-from-svn-sourcecode-webpage
- License text moved to bottom of main page
- Rest of the content should be really self explanatory and evident by checking out the screenshots, the 'live multiplayer and single player' game mode text on index.html page.

Perhaps of controversay a good bit of the about/features content is to describe and differentiate TripleA from other game clones, with some really notable features. At the same time the feature list sounds like a resume, "includes sound". So it's a bit tough to rewrite this content and boil it down, and eventually we have most of it or it should be pretty self-evident from the rest of the website.

With that there is nothing left on the about page, I was planning to rename it 'features' once the history content was removed and license moved, but makes sense to prune it altogether after those two updates.