Move win32 download to a link
There are more than expected windows 32 bit downloads than we would expect. To discourage users from choosing the wrong windows distro, the win32 download button is converted here to a link to "hide it". This will hopefully make it easier to pick the right distribution version, and if someone is not paying a lot of attention and is clicking the first thing that they see that says "windows", then they'll click the 64 bit windows DL now instead of 'accidently' getting the 32 bit.