Remove PDF version of rulebooks (#401)
This leaves only the HTML versions. It is therefore less to 'fix'
(there is still work needed). Of biggest note, anything that was
out of sync between PDF and HTML, is no longer an issue (the HTML
version would be the most up to date).

Helps address part of #347, reduces the size of the problem noted in #347.