Tactical Guide - restore images
The tactical guide was quickly (poorly) ported from PDF
to markdown. This update provides desperately needed
updates to restore the images that were previously present
in the PDF.

This update applies some formatting, eg: figure text
is italics and headers are bolder. More could be done,
the markdown version still does not look nearly as good
as the PDF version.

There is a PDF version of the tactical guide here:
https://www.moddb.com/downloads/start/39998

(google for 'TripleA-RuleBook-pdf')